Added a lobby and a simple 2D networking game
Added a lobby which allows one user to create a lobby as server and
other users to join that lobby. After everyone prepares, game will
start.
Added a simple 2D networking game which allows users to use keyboard(A,
D and Space) to control their characters to play with others.
Added transitions between all the scenes.